requirements-all.txt
transformers==4.30.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.3.0
langchain==0.0.335
PyPDF2==3.0.1
nltk==3.8.1
numpy==1.24.3
tqdm==4.66.1
python-dotenv==1.0.0
pymongo==4.6.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2